pred-maintenance/app/frontend/requirements.txt
streamlit==1.2.0
streamlit-option-menu==0.3.2
pillow
numpy